changes for swymedid in van and login swymed using email not username

+++ src/main/java/com/iemr/mmu/service/swymed/SwymedService.java
@@ -7,7 +7,7 @@ public interface SwymedService {
 	 String login(Long userid) throws SwymedException;
 	 
 	 String callUser(Long fromuserid,Long touserid) throws SwymedException;
-	 
-	 String callVan(Long fromuserid,Long vanid) throws SwymedException;
+
+	String callVan(Long fromuserid, Integer vanid) throws SwymedException;
 
 }
